backend/pyproject.toml
[project]
name = "ai-news-radio"
version = "0.1.0"
description = "AI-powered news radio pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "celery[redis]>=5.4",
    "httpx>=0.28",
    "anthropic>=0.43",
    "openai>=1.60",
    "google-generativeai>=0.8",
    "google-genai>=1.0",
    "mcp>=1.18.0",
    "pydantic-settings>=2.7",
    "beautifulsoup4>=4.12",
    "youtube-transcript-api>=0.6",
    "pypdf>=5.0",
    "python-pptx>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "aiosqlite>=0.20",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["B008"]  # Allow Depends() in FastAPI default arguments

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
